Results: Measurable improvement for JS -> Lisp Performance in multiple cases. 
All tests are based on 10,000 iterations 
| Test              | Pre-PR        | Post-PR  | % Improvement
| ------------- |:-------------:| ---------:| ------:|
| T1                | 775ms         |   356ms  |  54% |
| T2                | 45ms           |   8ms      | 82%  |
| T3                | 70ms           |   27ms    | 61%  |
| T4                | 32ms           |   22ms    | 31%  |

T1: (with-temp-buffer (insert "foo")) and JS equivalent
T2: (conat "foo" "bar") and JS equivalent 
T3: (intern "foo") and JS equivalent
T4: (insert "Hello World") and JS equivalent 

Lessons Learned:
* Using WeakRef caches is a powerful tool in the cases of tight loops calling lisp functions with similar arguments. Previously, when calling `lisp.insert("Hello")`, we would allocate a lisp string for every allocation, despite the fact that string does not change. Now we have a cached value that will only allocate once. 
* The bare minimum cost of a bridge cross between JS -> C/Rust is very small compared to the cost of allocating lisp or v8 JS objects. 
* `make_proxy!` was very slow, due to the fact we were creating an object template every time. We now have a global value that we save between calls. 
* Creating a lisp lambda from a JS lambda was slow. We now have a helper function to reduce the number of allocations and bridge crosses in creating lambda functions.